optimization: use fp16 for storage of low precision entities (images)
Most calculation is still done in fp64. From a basic test on a small dataset
I'm seeing a ~3x reduction in output data. More improvements in the pipeline

pre-reduction:
2.4G    output/J310/bin
2.6G    output/J310/CC
6.0M    output/J310/diff_maps
10G     output/J310/distances
7.1G    output/J310/ELConc50
34M     output/J310/pd_data.pkl
1.5K    output/J310/postproc
19G     output/J310/psi_analysis
146K    output/J310/topos
2.4G    output/J310/traj
43G     total

pass1: float16 images in storage
2.4G    output/J310/bin
666M    output/J310/CC
6.0M    output/J310/diff_maps
2.5G    output/J310/distances
1.8G    output/J310/ELConc50
34M     output/J310/pd_data.pkl
1.5K    output/J310/postproc
4.7G    output/J310/psi_analysis
146K    output/J310/topos
2.4G    output/J310/traj
15G     total

+++ ManifoldEM/CC/ComputeMeasureEdgeAll.py
@@ -37,8 +37,8 @@ def HOGOpticalFlowPy(flowVec):
 
         tempH = []
         for d in range(0, VxStackDim):
-            gx = flowVec['Vx'][:, :, d]
-            gy = flowVec['Vy'][:, :, d]
+            gx = flowVec['Vx'][:, :, d].astype(np.float64)
+            gy = flowVec['Vy'][:, :, d].astype(np.float64)
 
             tH = histogram_from_gradients(gx,
                                           gy,
@@ -56,8 +56,8 @@ def HOGOpticalFlowPy(flowVec):
             H = np.moveaxis(H, 0, -1)
     else:
 
-        gx = flowVec['Vx']
-        gy = flowVec['Vy']
+        gx = flowVec['Vx'].astype(np.float64)
+        gy = flowVec['Vy'].astype(np.float64)
         H = histogram_from_gradients(gx,
                                      gy,
                                      cell_size=cell_size,
